deleting fluent browser/element actions classes
