+++ src/main/java/com/shaft/driver/SHAFT.java
@@ -7,11 +7,11 @@ import com.shaft.cli.TerminalActions;
 import com.shaft.db.DatabaseActions;
 import com.shaft.driver.internal.DriverFactoryHelper;
 import com.shaft.driver.internal.WizardHelpers;
-import com.shaft.gui.browser.internal.FluentBrowserActions;
+import com.shaft.gui.browser.BrowserActions;
 import com.shaft.gui.element.AlertActions;
+import com.shaft.gui.element.ElementActions;
 import com.shaft.gui.element.SikuliActions;
 import com.shaft.gui.element.TouchActions;
-import com.shaft.gui.element.internal.FluentElementActions;
 import com.shaft.listeners.internal.WebDriverListener;
 import com.shaft.tools.io.ExcelFileManager;
 import com.shaft.tools.io.JSONFileManager;
@@ -21,7 +21,6 @@ import com.shaft.tools.io.internal.ReportManagerHelper;
 import com.shaft.validation.internal.RestValidationsBuilder;
 import io.appium.java_client.android.AndroidDriver;
 import io.appium.java_client.ios.IOSDriver;
-import io.restassured.config.RestAssuredConfig;
 import io.restassured.response.Response;
 import org.openqa.selenium.MutableCapabilities;
 import org.openqa.selenium.support.events.EventFiringDecorator;
@@ -63,16 +62,16 @@ public class SHAFT {
                 DriverFactoryHelper.closeDriver();
             }
 
-            public FluentElementActions element() {
-                return FluentElementActions.getInstance();
+            public ElementActions element() {
+                return ElementActions.getInstance();
             }
 
             public TouchActions touch() {
                 return new TouchActions();
             }
 
-            public FluentBrowserActions browser() {
-                return FluentBrowserActions.getInstance();
+            public BrowserActions browser() {
+                return BrowserActions.getInstance();
             }
 
             public AlertActions alert() {
